Create zpy-conda-env.yml
switched to conda deps
This is the --from-history export, it is cross platform
Will upload a version specific env maybe later. It is not needed in my opinion since the chances of thigns silently breaking with version updates are very low in these repos, functions might/will be deprecated and one will see the appropriate messages but silently breaking is unlikely to happen. But this is just my opinion.